These changes are mostly related to a new logic in TokenGen, affecting hyphens and eol conditions.
Created new functions break_hyphens and keep_hyphens in TokenGen. The
first of these is to be used for OCReval. The second is intended to be
used in the FormIndexer. Changed AccEval to use break_hyphens. Also
(not visible as a change in github), changed AccEval to check for match
with lowercase version of the word, because lexicon only includes lc
words.